When deleting an user that does not exit, return 404 and ResourceNotFoundExcepction

+++ src/main/java/com/desouza/client/service/ClientService.java
@@ -56,7 +56,11 @@ public class ClientService {
 
     @Transactional(propagation = Propagation.SUPPORTS)
     public void delete(Long id) {
-        repository.deleteById(id);
+        if (!repository.existsById(id)) {
+            throw new ResourceNotFoundException("Client with ID " + id + " not found");
+        } else {
+            repository.deleteById(id);
+        }
     }
 
     public void dtoToEntity(ClientDTO dto, Client entity) {
